+++ handler/routes.go
@@ -84,7 +84,18 @@ func Home(w http.ResponseWriter, r *http.Request, v *middleware.ValueMap) {
 	}
 
 	model := pagemodel.Home{
-		Menu: header,
+		Menu:     header,
+		ShowCase: []pagemodel.MeatModel{},
+	}
+
+	meats, err := db.GetMeatsPaging(8, 1)
+	if err != nil {
+		v.Set("warning", "Home: unable to get showcase meat >> "+err.Error())
+	} else {
+		for i := 0; i < len(meats); i++ {
+			meat := GetMeatModel(meats[i])
+			model.ShowCase = append(model.ShowCase, meat)
+		}
 	}
 
 	v.Set("next", false)
@@ -317,21 +328,6 @@ func ProductSearch(w http.ResponseWriter, r *http.Request, v *middleware.ValueMa
 	}
 
 	t.ExecuteTemplate(w, "product.html", model)
-}
-
-func GetMeatModel(meat dbmodel.Meat) pagemodel.MeatModel {
-	return pagemodel.MeatModel{
-		ID:          meat.ID.Hex(),
-		Pic:         "/image/meat_" + meat.ID.Hex() + meat.ImageExtension,
-		ProName:     meat.Name,
-		Type:        meat.Type,
-		Grade:       meat.Grade,
-		Description: meat.Description,
-		Price:       meat.Price,
-		Expire:      meat.Expire.Format(dbmodel.TimeFormat),
-		Quantity:    meat.Quantity,
-		Total:       meat.Price * float64(meat.Quantity),
-	}
 }
 
 func ProductDetail(w http.ResponseWriter, r *http.Request, v *middleware.ValueMap) {
