+++ model/pagemodel/home.go
@@ -2,4 +2,5 @@ package pagemodel
 
 type Home struct {
 	Menu
+	ShowCase []MeatModel
 }
