+++ template/layout_meat.html
@@ -1,0 +1,40 @@
+{{define "layout_meat"}}
+<div class="col-sm-6 col-md-4 col-lg-3 p-b-50">
+    <!-- Block2 -->
+    <div class="block2">
+        <div class="block2-img wrap-pic-w of-hidden pos-relative">
+            <img src="{{.Pic}}" alt="IMG-PRODUCT">
+
+            <div class="block2-overlay trans-0-4">
+                <a href="#" class="block2-btn-addwishlist hov-pointer trans-0-4">
+                    <i class="icon-wishlist icon_heart_alt" aria-hidden="true"></i>
+                    <i class="icon-wishlist icon_heart dis-none" aria-hidden="true"></i>
+                </a>
+
+                <div class="block2-btn-addcart w-size1 trans-0-4">
+                    <!-- Button -->
+                    <button class="flex-c-m size1 bg4 bo-rad-23 hov1 s-text1 trans-0-4 addCart" value="{{.ID}}">
+                        Add to Cart
+                    </button>
+                </div>
+            </div>
+        </div>
+
+        <div class="block2-txt p-t-20">
+            <a href="/product-detail/{{.ID}}" class="block2-name dis-block s-text3 p-b-5">
+                {{.ProName}} ({{.Type}})
+            </a>
+
+            <span class="block2-price m-text6 p-r-5">
+                ${{printf "%.2f" .Price}}
+            </span>
+        </div>
+    </div>
+</div>
+
+<!-- 
+in div class before img
+    block2-labelnew = new
+    block2-labelsale = sale
+     -->
+{{end}}
