+++ template/product.html
@@ -142,38 +142,7 @@
 					<div class="row">
 
 						{{range .Meats}}
-						<div class="col-sm-12 col-md-6 col-lg-4 p-b-50">
-							<!-- Block2 -->
-							<div class="block2">
-								<div class="block2-img wrap-pic-w of-hidden pos-relative block2-labelnew" id="fish">
-									<img src="{{.Pic}}" alt="IMG-PRODUCT">
-
-									<div class="block2-overlay trans-0-4">
-										<a href="#" class="block2-btn-addwishlist hov-pointer trans-0-4">
-											<i class="icon-wishlist icon_heart_alt" aria-hidden="true"></i>
-											<i class="icon-wishlist icon_heart dis-none" aria-hidden="true"></i>
-										</a>
-
-										<div class="block2-btn-addcart w-size1 trans-0-4">
-											<!-- Button -->
-											<button class="flex-c-m size1 bg4 bo-rad-23 hov1 s-text1 trans-0-4 addCart" value="{{.ID}}">
-												Add to Cart
-											</button>
-										</div>
-									</div>
-								</div>
-
-								<div class="block2-txt p-t-20">
-									<a href="/product-detail/{{.ID}}/" class="block2-name dis-block s-text3 p-b-5" id="product-name">
-										{{.ProName}} ({{.Type}})
-									</a>
-
-									<span class="block2-price m-text6 p-r-5" id="product-price">
-										${{printf "%.2f" .Price}}
-									</span>
-								</div>
-							</div>
-						</div>
+							{{template "layout_meat" .}}
 						{{end}}
 
 					</div>
@@ -282,7 +251,7 @@
 	</script>
 	<!--===============================================================================================-->
 	<script src="/static/js/main.js"></script>
-	<script src="../static/js/addCart.js"></script>
+	<script src="/static/js/addCart.js"></script>
 
 </body>
